Refactor migrations, factories, and seeders for production readiness
- Migrations: Remove comments, optimize data types for PostgreSQL (unsignedSmallInteger, timestampTz, text), add performance indexes
- Merge fix_push_subscriptions migration into create_push_subscriptions
- Factories: Remove comments, add placeholder images (placehold.co), optimize fake data generation
- Seeders: Update subjects for 2-4 courses, expand technologies list (PHP, Java, C#, JS/TS, Python, Docker, Cloud), refactor all seeders
- Add ProductionSeeder for essential data (roles, subjects, technologies, categories, admin)
- Update entrypoint.sh: migrate:fresh with ProductionSeeder for production, full db:seed for development

+++ database/migrations/2025_01_21_115519_create_subjects_table.php
@@ -6,25 +6,19 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('subjects', function (Blueprint $table) {
             $table->ulid('id')->primary();
             $table->string('slug', 72)->unique();
             $table->string('name', 128);
-            $table->integer('course_number')->nullable();
+            $table->unsignedSmallInteger('course_number')->nullable();
             $table->text('description')->nullable();
             $table->string('image', 2048)->nullable();
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('subjects');
