+++ database/migrations/2025_01_21_120001_create_categories_table.php
@@ -6,25 +6,19 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('categories', function (Blueprint $table) {
             $table->ulid('id')->primary();
             $table->string('name', 32);
-            $table->integer('freezing_period')->nullable();
-            $table->integer('course_number')->nullable();
-            $table->integer('period');
+            $table->unsignedSmallInteger('freezing_period')->nullable();
+            $table->unsignedSmallInteger('course_number')->nullable();
+            $table->unsignedSmallInteger('period');
             $table->jsonb('attachments')->nullable();
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('categories');
